Completely refactored source tree
